+++ src/ralph_assets/models_util.py
@@ -15,15 +15,6 @@ from django.db import models
 from lck.django.choices import Choices
 
 
-class SavingUser(models.Model):
-    class Meta:
-        abstract = True
-
-    def save(self, user=None, *args, **kwargs):
-        self.saving_user = user
-        return super(SavingUser, self).save(*args, **kwargs)
-
-
 class ProblemSeverity(Choices):
     _ = Choices.Choice
     warning = _("Warning")