+++ src/ralph_assets/tests/functional/tests_history.py
@@ -15,7 +15,7 @@ from ralph_assets.models_assets import (
     SAVE_PRIORITY,
     AssetType
 )
-from ralph_assets.models_history import AssetHistoryChange
+from ralph_assets.history.models import History
 from ralph_assets.tests.utils.assets import (
     AssetCategoryFactory,
     AssetManufacturerFactory,
@@ -26,6 +26,7 @@ from ralph_assets.tests.utils.assets import (
     ServiceCatalogFactory,
     WarehouseFactory,
 )
+from ralph_assets.tests.utils.sam import LicenceFactory
 from ralph.business.models import Venture
 from ralph.discovery.models_device import Device, DeviceType
 from ralph.ui.tests.global_utils import login_as_su
@@ -41,6 +42,7 @@ class HistoryAssetsView(TestCase):
             manufacturer=self.manufacturer,
             category=self.category,
         )
+        self.licences = [LicenceFactory() for _ in range(3)]
         self.warehouse = WarehouseFactory()
         self.asset_params = {
             'asset': True,  # Button name
@@ -74,6 +76,7 @@ class HistoryAssetsView(TestCase):
             'license_type': LicenseType.oem.id,
             'date_of_last_inventory': '2012-11-08',
             'last_logged_user': 'ralph',
+            'licences': '|'.join([str(lic.pk) for lic in self.licences]),
         }
         self.dc_asset_params = self.asset_params.copy()
         self.dc_asset_params.update({
@@ -91,12 +94,17 @@ class HistoryAssetsView(TestCase):
         self.add_bo_device_asset()
         self.edit_bo_device_asset()
 
+    def convert_to_list_str(self, pipes):
+            if pipes == '':
+                return '[]'
+            return '[' + ', '.join(pipes.split('|')) + ']'
+
     def add_bo_device_asset(self):
         """Test check adding Asset into backoffice through the form UI"""
         url = '/assets/back_office/add/device/'
         attrs = self.bo_asset_params
-        request = self.client.post(url, attrs)
-        self.assertEqual(request.status_code, 302)
+        request = self.client.post(url, attrs, follow=True)
+        self.assertEqual(request.status_code, 200)
 
     def edit_bo_device_asset(self):
         """Test checks asset edition through the form UI"""
@@ -106,14 +114,12 @@ class HistoryAssetsView(TestCase):
             self.bo_asset_params.items() + self.asset_change_params.items()
         )
         attrs.update({'purpose': 2})
-        request = self.client.post(url, attrs)
-        self.assertEqual(request.status_code, 302)
+        request = self.client.post(url, attrs, follow=True)
+        self.assertEqual(request.status_code, 200)
 
     def test_change_status(self):
         """Test check the recording Asset status change in asset history"""
-        asset_history = AssetHistoryChange.objects.get(
-            asset=self.asset, field_name='status'
-        )
+        asset_history = self.asset.get_history(field_name='status')[0]
         self.assertListEqual(
             [asset_history.old_value, asset_history.new_value],
             [AssetStatus.new.name, AssetStatus.damaged.name]
@@ -121,13 +127,30 @@ class HistoryAssetsView(TestCase):
 
     def test_change_barcode(self):
         """Test check the recording Asset barcode change in asset history"""
-        asset_history = AssetHistoryChange.objects.filter(
-            asset=self.asset, field_name='barcode'
-        )
+        asset_history = self.asset.get_history(field_name='barcode')[0]
         self.assertListEqual(
-            [asset_history[0].old_value, asset_history[0].new_value],
+            [asset_history.old_value, asset_history.new_value],
             [self.asset_params['barcode'], self.asset_change_params['barcode']]
         )
+
+    def test_change_licences(self):
+        """Test check the recording Asset licence set change
+        in asset history"""
+        asset_history = self.asset.get_history(field_name='licence_set')[0]
+        self.assertListEqual(
+            [asset_history.old_value, asset_history.new_value],
+            [
+                self.convert_to_list_str(self.asset_params['licences']),
+                self.convert_to_list_str(self.asset_change_params['licences'])
+            ]
+        )
+
+        for licence in self.licences:
+            licence_history = licence.get_history(field_name='assets')[0]
+            self.assertEqual(licence_history.old_value, '[]')
+            self.assertEqual(
+                licence_history.new_value, '[{}]'.format(self.asset.id)
+            )
 
     def test_change_required_support(self):
         asset = BOAssetFactory()
@@ -143,10 +166,7 @@ class HistoryAssetsView(TestCase):
             'asset': True,
         })
         response = self.client.post(url, update_dict)
-        url = reverse('device_history', kwargs={
-            'mode': 'back_office',
-            'asset_id': asset.id,
-        })
+        url = History.get_history_url_for_object(asset)
         response = self.client.get(url)
         self.assertContains(response, 'required_support')